+++ src/components/Checkbox/Checkbox.stories.js
@@ -11,8 +11,17 @@ export default {
 const Template = (args, { argTypes }) => ({
   components: { PCheckbox },
   props: Object.keys(argTypes),
-  template:
-    "<PCheckbox v-bind='$props' v-model='optIn' :errors='optIn === false && [`Required`] || []' /><div>some texts</div>",
+  template: `
+    <div>
+      <PCheckbox
+        v-bind='$props'
+        v-model='optIn'
+        :errors="optIn === false && ['Required'] || []"
+      >
+        <div>Photon is awesome</div>
+      </PCheckbox>
+      </div>
+    `,
   data() {
     return {
       optIn: false,
@@ -21,6 +30,4 @@ const Template = (args, { argTypes }) => ({
 });
 
 export const Checkbox = Template.bind({});
-Checkbox.args = {
-  label: 'Photon is awesome',
-};
+Checkbox.args = {};